Deployment adjustments and refactoring
- Added __init__.py file to packages to have them recognized as python-packages by deployment platform.

- Factored models (all http entities) out of api.py file in its own file in the service package.
