+++ vedic-agent-api/service/__init__.py
@@ -1,0 +1,1 @@
+# Empty file to mark the service directory as a Python package
