+++ vedic-agent-api/service/vedic_retrieval_api.py
@@ -16,6 +16,7 @@ from contextlib import asynccontextmanager
 from fastapi.middleware.cors import CORSMiddleware
 
 from service.VedicRetriever import VedicRetriever
+from service.models import FeedbackRequest, RetrievalResponse
 
 warnings.filterwarnings("ignore", category=UserWarning, module="urllib3")
 logging.basicConfig(level=logging.INFO, format="%(asctime)s - %(levelname)s - %(message)s")
@@ -26,6 +27,7 @@ load_dotenv()
 
 app = FastAPI(title="VedicSage Retrieval API",
               description="Retrieve relevant Vedic verses based on semantic similarity")
+
 # Add CORS middleware
 app.add_middleware(
     CORSMiddleware,
@@ -34,15 +36,6 @@ app.add_middleware(
     allow_methods=["*"],  # Allow all methods (GET, POST, etc.)
     allow_headers=["*"],  # Allow all headers
 )
-class RetrievalResult(BaseModel):
-    verse: str
-    source: str
-    score: float
-
-
-class RetrievalResponse(BaseModel):
-    query: str
-    results: List[RetrievalResult]
 
 
 def find_file(base_path: str, filename: str) -> str:
@@ -61,65 +54,6 @@ def find_file(base_path: str, filename: str) -> str:
             logger.info(f"Found {filename} at: {file_path}")
             return file_path
     raise FileNotFoundError(f"File '{filename}' not found in common directories: {search_dirs}")
-
-
-# class VedicRetriever:
-#     def __init__(self, index_path: str, metadata_path: str, model_name: str = "all-MiniLM-L6-v2"):
-#         try:
-#             if not Path(index_path).exists():
-#                 logger.warning(f"FAISS index not found at {index_path}, searching common directories...")
-#                 index_path = find_file(os.path.dirname(index_path), "verse_index.faiss")
-#             self.index = faiss.read_index(index_path)
-#             logger.info(f"Loaded FAISS index from {index_path}")
-#
-#             if not Path(metadata_path).exists():
-#                 logger.warning(f"Metadata CSV not found at {metadata_path}, searching common directories...")
-#                 metadata_path = find_file(os.path.dirname(metadata_path), "verses_metadata.csv")
-#             self.df = pd.read_csv(metadata_path, encoding='utf-8')
-#             # old column format
-#             # required_columns = ['id', 'book', 'chapter', 'verse', 'text_en']
-#             required_columns = ['verse_id', 'text', 'source']
-#             if not all(col in self.df.columns for col in required_columns):
-#                 logger.error(f"Metadata missing required columns: {required_columns}")
-#                 raise ValueError(f"Metadata missing required columns")
-#             logger.info(f"Loaded {len(self.df)} verses from {metadata_path}")
-#
-#             logger.info(f"Loading Sentence-BERT model: {model_name}")
-#             self.model = SentenceTransformer(model_name)
-#         except Exception as e:
-#             logger.error(f"Error initializing retriever: {e}")
-#             raise
-#
-#     def retrieve(self, query: str, k: int = 5) -> List[Dict]:
-#         try:
-#             query_emb = self.model.encode([query]).astype(np.float32)
-#             distances, indices = self.index.search(query_emb, k)
-#             results = []
-#             for dist, idx in zip(distances[0], indices[0]):
-#                 if idx < len(self.df):
-#                     verse = self.df.iloc[idx]
-#                     score = 1 / (1 + dist)
-#                     # other format:
-#                     # source = f"{verse['book']} {verse['chapter']}.{verse['verse']}"
-#                     # results.append({
-#                     #     "verse": verse['text_en'],
-#                     #     "source": source,
-#                     #     "score": float(score)
-#                     # })
-#
-#                     results.append({
-#                         "verse": verse["text"],
-#                         "source": verse["source"],
-#                         "score": float(score)
-#                     })
-#                 else:
-#                     logger.warning(f"Invalid index {idx} returned by FAISS")
-#             return results
-#         except Exception as e:
-#             logger.error(f"Error retrieving for query '{query}': {e}")
-#             raise
-#
-
 
 
 
@@ -214,13 +148,9 @@ async def answer_question(query: str, k: int = 5):
         # raise HTTPException(status_code=500, detail="Failed to generate answer")
 
 
-# ## FEEDBACK
+### FEEDBACK
 from pydantic import BaseModel
 
-class FeedbackRequest(BaseModel):
-    query: str
-    verse_id: str
-    is_relevant: bool  # True for relevant, False for not relevant
 
 @app.post("/feedback")
 async def submit_feedback(feedback: FeedbackRequest):
@@ -254,8 +184,6 @@ if __name__ == "__main__":
     import os
 
     logger.info(f"Working directory: {os.getcwd()}")
-
-
 
     port = int(os.environ.get("PORT", 8000))
     uvicorn.run(app, host="0.0.0.0", port=port)
